+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- New config editor page
+    - Added right click option to jump to config
+
 ## [0.5.1] 2019-05-04
 
 ### Added